+++ src/config/recommended.ts
@@ -9,6 +9,7 @@ module.exports = {
 		'element-permitted-content': 'error',
 		'element-permitted-occurrences': 'error',
 		'element-permitted-order': 'error',
+		'img-req-alt': 'error',
 		'no-dup-attr': 'error',
 		'no-dup-class': 'error',
 		'no-dup-id': 'error',