+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 - new rule `element-permitted-occurrences` validating elements with content
   models limiting the number of times child elements can be used.
 
+### Bugfixes
+
+- the parser now resets the DOM tree before starting, fixes issue when running
+  the same parser instance on multiple sources.
+
 ## 0.3.0 (2017-10-12)
 
 ### Features